Modify Detekt configuration to be more strict
Generally enable all rules that are not obviously wrong-headed (for
instance, leave `CommentOverPrivateFunction` off because I don't think
all private constructs can be made self-documenting).

Also change the "exclude test sources" lines to only exclude the default
JVM test sources; this is not an Android or KMP project, so all the
extra exclusions are just unnecessary noise for us.