Set up skeleton of core module
Add a second module, called `core`.  This is where we will define the
central `Exercise` interface that will be shared across all problem
families (like Advent of Code or Project Euler) and accessed through the
CLI.

At this initial step, we simply create the module itself and delegate
to it the Hello World implementation from the CLI.  Which is, of course,
a ridiculous level of indirection for the current behavior, but it lets
us establish structure before we worry about more meaningful concerns.